draw indicators with chart
